allow easier custom audio device setup

+++ files/data/config/systems/allwinner_h3_tv_box.sh
@@ -4,8 +4,13 @@ if [ "$?" = "0" ]; then
   ln -sf /opt/mali-sunxi-fbdev-armv7l /opt/libgl
   ln -sf /opt/gl4es-armv7l /opt/gl4es
   cp /data/config/x11/xorg.conf-sunxi /etc/X11/xorg.conf.d/xorg.conf
-  cp /data/config/qjackctl/QjackCtl.conf-h3 /data/config/qjackctl/QjackCtl.conf
-  ( sleep 15; AUDIO_DEVICE=`aplay -l | grep "H3 Audio Codec" | awk '{print $2}' | sed 's,:,,g'`; if [ "$AUDIO_DEVICE" != "" ]; then amixer -c ${AUDIO_DEVICE} set 'Line Out' 31 ; amixer -c ${AUDIO_DEVICE} set DAC 63 ; fi ) &
+  # check if a custom audio setup exists and use it in that case
+  if [ -f /data/config/custom/audio-setup.sh ]; then
+    . /data/config/custom/audio-setup.sh
+  else
+    cp /data/config/qjackctl/QjackCtl.conf-h3 /data/config/qjackctl/QjackCtl.conf
+    ( sleep 15; AUDIO_DEVICE=`aplay -l | grep "H3 Audio Codec" | awk '{print $2}' | sed 's,:,,g'`; if [ "$AUDIO_DEVICE" != "" ]; then amixer -c ${AUDIO_DEVICE} set 'Line Out' 31 ; amixer -c ${AUDIO_DEVICE} set DAC 63 ; fi ) &
+  fi
   echo "SYSTEM_MODEL=h3" > /data/config/info.txt
   echo "SYSTEM_MODEL_DETAILED=allwinner_h3_tv_box" >> /data/config/info.txt
   # start vcvrack v0 with realtime scheduling priority - might result in system hangs
