+++ files/data/config/systems/amlogic_s905_w_x_tv_box.sh
@@ -9,8 +9,13 @@ if [ "$?" = "0" ]; then
     ln -sf /opt/gl4es-armv7l /opt/gl4es
   fi
   cp /data/config/x11/xorg.conf-s905 /etc/X11/xorg.conf.d/xorg.conf
-  cp /data/config/qjackctl/QjackCtl.conf-pcm2704 /data/config/qjackctl/QjackCtl.conf
-  ( sleep 15; AUDIO_DEVICE=`aplay -l | grep "DAC \[USB AUDIO    DAC\]" | awk '{print $2}' | sed 's,:,,g'`; if [ "$AUDIO_DEVICE" != "" ]; then amixer -c ${AUDIO_DEVICE} set PCM 64 ; fi ) &
+  # check if a custom audio setup exists and use it in that case
+  if [ -f /data/config/custom/audio-setup.sh ]; then
+    . /data/config/custom/audio-setup.sh
+  else
+    cp /data/config/qjackctl/QjackCtl.conf-pcm2704 /data/config/qjackctl/QjackCtl.conf
+    ( sleep 15; AUDIO_DEVICE=`aplay -l | grep "DAC \[USB AUDIO    DAC\]" | awk '{print $2}' | sed 's,:,,g'`; if [ "$AUDIO_DEVICE" != "" ]; then amixer -c ${AUDIO_DEVICE} set PCM 64 ; fi ) &
+  fi
   echo "SYSTEM_MODEL=s905" > /data/config/info.txt
   echo "SYSTEM_MODEL_DETAILED=amlogic_s905_w_x_tv_box" >> /data/config/info.txt
   # start vcvrack v0 with realtime scheduling priority - might result in system hangs
