+++ files/data/config/systems/bananapi_m1.sh
@@ -4,8 +4,13 @@ if [ "$?" = "0" ]; then
   ln -sf /opt/mali-sunxi-fbdev-armv7l /opt/libgl
   ln -sf /opt/gl4es-armv7l /opt/gl4es
   cp /data/config/x11/xorg.conf-sunxi /etc/X11/xorg.conf.d/xorg.conf
-  cp /data/config/qjackctl/QjackCtl.conf-pcm2704 /data/config/qjackctl/QjackCtl.conf
-  ( sleep 15; AUDIO_DEVICE=`aplay -l | grep "DAC \[USB AUDIO    DAC\]" | awk '{print $2}' | sed 's,:,,g'`; if [ "$AUDIO_DEVICE" != "" ]; then amixer -c ${AUDIO_DEVICE} set PCM 64 ; fi ) &
+  # check if a custom audio setup exists and use it in that case
+  if [ -f /data/config/custom/audio-setup.sh ]; then
+    . /data/config/custom/audio-setup.sh
+  else
+    cp /data/config/qjackctl/QjackCtl.conf-pcm2704 /data/config/qjackctl/QjackCtl.conf
+    ( sleep 15; AUDIO_DEVICE=`aplay -l | grep "DAC \[USB AUDIO    DAC\]" | awk '{print $2}' | sed 's,:,,g'`; if [ "$AUDIO_DEVICE" != "" ]; then amixer -c ${AUDIO_DEVICE} set PCM 64 ; fi ) &
+  fi
   echo "SYSTEM_MODEL=bananapi_m1" > /data/config/info.txt
   echo "SYSTEM_MODEL_DETAILED=bananapi_m1" >> /data/config/info.txt
   # start vcvrack v0 with realtime scheduling priority - might result in system hangs
