+++ files/data/config/systems/raspberry_pi_3_b.sh
@@ -9,8 +9,13 @@ if [ "$?" = "0" ]; then
     ln -sf /dev/null /opt/gl4es
   fi
   cp /data/config/x11/xorg.conf-raspberrypi /etc/X11/xorg.conf.d/xorg.conf
-  cp /data/config/qjackctl/QjackCtl.conf-raspberrypi /data/config/qjackctl/QjackCtl.conf
-  ( sleep 15; AUDIO_DEVICE=`aplay -l | grep "bcm2835 ALSA \[bcm2835 ALSA\]" | awk '{print $2}' | sed 's,:,,g'`; if [ "$AUDIO_DEVICE" != "" ]; then amixer -c ${AUDIO_DEVICE} set PCM 0 ; fi ) &
+  # check if a custom audio setup exists and use it in that case
+  if [ -f /data/config/custom/audio-setup.sh ]; then
+    . /data/config/custom/audio-setup.sh
+  else
+    cp /data/config/qjackctl/QjackCtl.conf-raspberrypi /data/config/qjackctl/QjackCtl.conf
+    ( sleep 15; AUDIO_DEVICE=`aplay -l | grep "bcm2835 ALSA \[bcm2835 ALSA\]" | awk '{print $2}' | sed 's,:,,g'`; if [ "$AUDIO_DEVICE" != "" ]; then amixer -c ${AUDIO_DEVICE} set PCM 0 ; fi ) &
+  fi
   echo "SYSTEM_MODEL=raspberrypi" > /data/config/info.txt
   echo "SYSTEM_MODEL_DETAILED=raspberrypi_3_b" >> /data/config/info.txt
   # start vcvrack v0 with realtime scheduling priority - might result in system hangs
